add stream for register and resource register
