+++ Altinn.Authorization.sln
@@ -99,8 +99,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Altinn.Authorization.Integr
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{3C712914-2427-4491-BC8D-5A7AC45C6D82}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Authorization.Integration.Register", "src\libs\Altinn.Authorization.Integration\src\Altinn.Authorization.Integration.Register\Altinn.Authorization.Integration.Register.csproj", "{E29CE48E-2901-4E9A-957D-E1D367F7D7F2}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Altinn.Authorization.Host", "Altinn.Authorization.Host", "{70F3B88D-68BD-4B7A-8A19-D8F796522595}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{FE84C40C-0A57-497A-9E9D-7E446B730239}"
@@ -114,6 +112,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{DC82C1C6-C706-4785-80D0-4D5D4F6BA620}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Authorization.Host.Lease.Tests", "src\libs\Altinn.Authorization.Host\test\Altinn.Authorization.Host.Lease.Tests\Altinn.Authorization.Host.Lease.Tests.csproj", "{6D460AF5-06D5-41BD-A612-FCC7209C989F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Authorization.Integration.Platform", "src\libs\Altinn.Authorization.Integration\src\Altinn.Authorization.Integration.Platform\Altinn.Authorization.Integration.Platform.csproj", "{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -212,10 +212,6 @@ Global
 		{67201B93-0606-4C34-8B67-8F693006837F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67201B93-0606-4C34-8B67-8F693006837F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67201B93-0606-4C34-8B67-8F693006837F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E29CE48E-2901-4E9A-957D-E1D367F7D7F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E29CE48E-2901-4E9A-957D-E1D367F7D7F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E29CE48E-2901-4E9A-957D-E1D367F7D7F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E29CE48E-2901-4E9A-957D-E1D367F7D7F2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{613F1E18-2B99-4466-B23E-240C1853A144}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{613F1E18-2B99-4466-B23E-240C1853A144}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{613F1E18-2B99-4466-B23E-240C1853A144}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -232,6 +228,10 @@ Global
 		{6D460AF5-06D5-41BD-A612-FCC7209C989F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D460AF5-06D5-41BD-A612-FCC7209C989F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D460AF5-06D5-41BD-A612-FCC7209C989F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{E9D041A5-2AB6-45FD-8D24-EF552025539E} = {2891B160-9E46-42E5-95FF-08523A3192EF}
@@ -281,7 +281,6 @@ Global
 		{67201B93-0606-4C34-8B67-8F693006837F} = {EF1C3A6D-0085-4BA9-BAED-B75938211F4E}
 		{840BDEEB-6600-42C8-A9FA-50B70A077454} = {505CF9AC-8997-4AB6-81DE-1B5EE2A75318}
 		{3C712914-2427-4491-BC8D-5A7AC45C6D82} = {840BDEEB-6600-42C8-A9FA-50B70A077454}
-		{E29CE48E-2901-4E9A-957D-E1D367F7D7F2} = {3C712914-2427-4491-BC8D-5A7AC45C6D82}
 		{70F3B88D-68BD-4B7A-8A19-D8F796522595} = {505CF9AC-8997-4AB6-81DE-1B5EE2A75318}
 		{FE84C40C-0A57-497A-9E9D-7E446B730239} = {70F3B88D-68BD-4B7A-8A19-D8F796522595}
 		{613F1E18-2B99-4466-B23E-240C1853A144} = {FE84C40C-0A57-497A-9E9D-7E446B730239}
@@ -289,5 +288,6 @@ Global
 		{DA0ECA50-BCE0-4067-8946-71B6A654BC8C} = {FE84C40C-0A57-497A-9E9D-7E446B730239}
 		{DC82C1C6-C706-4785-80D0-4D5D4F6BA620} = {70F3B88D-68BD-4B7A-8A19-D8F796522595}
 		{6D460AF5-06D5-41BD-A612-FCC7209C989F} = {DC82C1C6-C706-4785-80D0-4D5D4F6BA620}
+		{1067CBED-98BE-4E8E-AA7F-35138DF2E3F4} = {3C712914-2427-4491-BC8D-5A7AC45C6D82}
 	EndGlobalSection
 EndGlobal
