+++ src/apps/Altinn.Authorization/infra/at22.tfvars
@@ -1,0 +1,8 @@
+environment = "at22"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/cb4eb3bd-3ee3-4e48-ad7c-db31c12f5f64/"
+    namespace       = "default"
+    service_account = "altinn-authorization"
+  }
+]
