+++ src/apps/Altinn.Authorization/infra/at23.tfvars
@@ -1,0 +1,8 @@
+environment = "at23"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/68c4e940-5ccb-4825-9eb5-2bb429e242b5/"
+    namespace       = "default"
+    service_account = "altinn-authorization"
+  }
+]
