+++ src/apps/Altinn.Authorization/infra/at24.tfvars
@@ -1,0 +1,8 @@
+environment = "at24"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/4a142a74-1861-42c5-9014-36cc46ed71a0/"
+    namespace       = "default"
+    service_account = "altinn-authorization"
+  }
+]
