+++ src/apps/Altinn.Authorization/infra/prod.tfvars
@@ -1,0 +1,8 @@
+environment = "prod"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/f61fe162-dabf-4796-b3e3-694a0503ffa2/"
+    namespace       = "default"
+    service_account = "altinn-access-management"
+  }
+]
