+++ src/apps/Altinn.Authorization/infra/tt02.tfvars
@@ -1,0 +1,8 @@
+environment = "tt02"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/d118512b-ed4c-4387-8ef5-c893a1e7b788/"
+    namespace       = "default"
+    service_account = "altinn-authorization"
+  }
+]
