+++ src/apps/Altinn.Authorization/infra/yt01.tfvars
@@ -1,0 +1,8 @@
+environment = "yt01"
+aks_federation = [
+  {
+    issuer_url      = "https://norwayeast.oic.prod-aks.azure.com/cd0026d8-283b-4a55-9bfa-d0ef4a8ba21c/6c3cbd25-2f6e-4003-b073-a635b8d0a1b1/"
+    namespace       = "default"
+    service_account = "altinn-authorization"
+  }
+]
