+++ src/libs/Altinn.Authorization.Integration/src/Altinn.Authorization.Integration.Platform/Altinn.Authorization.Integration.Platform.csproj
@@ -7,5 +7,4 @@
         <PackageReference Include="Altinn.Common.AccessTokenClient" />
     </ItemGroup>
 
-
 </Project>
