+++ src/libs/Altinn.Authorization.Integration/src/Altinn.Authorization.Integration.Platform/Register/AltinnRegisterOptions.cs
@@ -1,7 +1,7 @@
-namespace Altinn.Authorization.Integration.Register.Options;
+namespace Altinn.Authorization.Integration.Platform.Register;
 
 /// <summary>
-/// Configuration options for Altinn Register integration.
+/// Configuration options for the Altinn Register integration.
 /// </summary>
 public class AltinnRegisterOptions
 {
@@ -13,10 +13,12 @@ public class AltinnRegisterOptions
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="AltinnRegisterOptions"/> class and applies
-    /// the provided configuration action to set up the options.
+    /// Initializes a new instance of the <see cref="AltinnRegisterOptions"/> class 
+    /// and applies the provided configuration action to set up the options.
     /// </summary>
-    /// <param name="configureOptions">A delegate to configure the <see cref="AltinnRegisterOptions"/> instance.</param>
+    /// <param name="configureOptions">
+    /// A delegate that configures the <see cref="AltinnRegisterOptions"/> instance.
+    /// </param>
     public AltinnRegisterOptions(Action<AltinnRegisterOptions> configureOptions)
     {
         configureOptions(this);
@@ -26,7 +28,7 @@ public class AltinnRegisterOptions
     /// Gets or sets the endpoint URL for the Altinn Register service.
     /// </summary>
     /// <remarks>
-    /// This URL is used to make requests to the Altinn Register API.
+    /// This URL is used for making requests to the Altinn Register API.
     /// </remarks>
     public Uri Endpoint { get; set; }
 }
