+++ src/libs/Altinn.Authorization.Integration/src/Altinn.Authorization.Integration.Platform/ResourceRegister/AltinnResourceRegisterOptions.cs
@@ -1,0 +1,32 @@
+namespace Altinn.Authorization.Integration.Platform.ResourceRegister;
+
+/// <summary>
+/// Configuration options for Altinn Resource Register integration.
+/// </summary>
+public class AltinnResourceRegisterOptions
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="AltinnResourceRegisterOptions"/> class.
+    /// </summary>
+    public AltinnResourceRegisterOptions()
+    {
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="AltinnResourceRegisterOptions"/> class and applies
+    /// the provided configuration action to set up the options.
+    /// </summary>
+    /// <param name="configureOptions">A delegate to configure the <see cref="AltinnResourceRegisterOptions"/> instance.</param>
+    public AltinnResourceRegisterOptions(Action<AltinnResourceRegisterOptions> configureOptions)
+    {
+        configureOptions(this);
+    }
+
+    /// <summary>
+    /// Gets or sets the endpoint URL for the Altinn Register service.
+    /// </summary>
+    /// <remarks>
+    /// This URL is used to make requests to the Altinn Register API.
+    /// </remarks>
+    public Uri Endpoint { get; set; }
+}
