--- src/libs/Altinn.Authorization.Integration/src/Altinn.Authorization.Integration.Register/Models/Paginated.cs
@@ -1,30 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Altinn.Authorization.Integration.Register.Models;
-
-/// <summary>
-/// A paginated <see cref="ListObject{T}"/>.
-/// </summary>
-/// <typeparam name="T">The item type.</typeparam>
-/// <param name="Links">Pagination links.</param>
-/// <param name="Items">The items.</param>
-public record Paginated<T>(
-    PaginatedLinks Links,
-    IEnumerable<T> Items)
-    : ListObject<T>(Items);
-
-/// <summary>
-/// Pagination links.
-/// </summary>
-/// <param name="Next">Link to the next page of items (if any).</param>
-public record PaginatedLinks(
-    string Next);
-
-/// <summary>
-/// A concrete list object.
-/// </summary>
-/// <typeparam name="T">The item type.</typeparam>
-/// <param name="Items">The items.</param>
-public record ListObject<T>(
-    [property: JsonPropertyName("data")]
-    IEnumerable<T> Items);
